refactor: Relocated classification models to models.classification
