+++ holocron/models/__init__.py
@@ -1,12 +1,2 @@
 from . import detection, segmentation
-from .darknet import *
-from .darknetv2 import *
-from .darknetv3 import *
-from .darknetv4 import *
-from .pyconvresnet import *
-from .repvgg import *
-from .res2net import *
-from .resnet import *
-from .rexnet import *
-from .sknet import *
-from .tridentnet import *
+from .classification import *
