+++ holocron/models/classification/darknetv3.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -10,11 +10,11 @@ import torch
 import torch.nn as nn
 
 from holocron.nn import DropBlock2d, GlobalAvgPool2d
+from holocron.nn.init import init_module
 
-from ..nn.init import init_module
-from .presets import IMAGENETTE
+from ..presets import IMAGENETTE
+from ..utils import conv_sequence, load_pretrained_params
 from .resnet import _ResBlock
-from .utils import conv_sequence, load_pretrained_params
 
 __all__ = ['DarknetV3', 'darknet53']
