+++ holocron/models/classification/pyconvresnet.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -9,9 +9,9 @@ from torch.nn import Module
 
 from holocron.nn import PyConv2d
 
-from .presets import IMAGENETTE
+from ..presets import IMAGENETTE
+from ..utils import conv_sequence, load_pretrained_params
 from .resnet import ResNet, _ResBlock
-from .utils import conv_sequence, load_pretrained_params
 
 __all__ = ['PyBottleneck', 'pyconv_resnet50', 'pyconvhg_resnet50']
