+++ holocron/models/classification/resnet.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -11,8 +11,8 @@ from torch import Tensor
 
 from holocron.nn import GlobalAvgPool2d, init
 
-from .presets import IMAGENET, IMAGENETTE
-from .utils import conv_sequence, load_pretrained_params
+from ..presets import IMAGENET, IMAGENETTE
+from ..utils import conv_sequence, load_pretrained_params
 
 __all__ = ['BasicBlock', 'Bottleneck', 'ResNet', 'resnet18', 'resnet34', 'resnet50', 'resnet101', 'resnet152',
            'resnext50_32x4d', 'resnext101_32x8d', 'resnet50d']
