+++ holocron/models/classification/sknet.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -10,9 +10,9 @@ import torch.nn as nn
 
 from holocron.nn import GlobalAvgPool2d
 
-from .presets import IMAGENETTE
+from ..presets import IMAGENETTE
+from ..utils import conv_sequence, load_pretrained_params
 from .resnet import ResNet, _ResBlock
-from .utils import conv_sequence, load_pretrained_params
 
 __all__ = ['SoftAttentionLayer', 'SKConv2d', 'SKBottleneck', 'sknet50', 'sknet101', 'sknet152']
