+++ holocron/models/classification/tridentnet.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -9,9 +9,9 @@ import torch
 import torch.nn as nn
 from torch.nn import functional as F
 
-from .presets import IMAGENETTE
+from ..presets import IMAGENETTE
+from ..utils import conv_sequence, load_pretrained_params
 from .resnet import ResNet, _ResBlock
-from .utils import conv_sequence, load_pretrained_params
 
 __all__ = ['Tridentneck', 'tridentnet50']
