+++ holocron/models/detection/yolo.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -13,8 +13,8 @@ from torchvision.ops.boxes import box_iou, nms
 
 from holocron.nn.init import init_module
 
-from ..darknet import DarknetBodyV1
-from ..darknet import default_cfgs as dark_cfgs
+from ..classification.darknet import DarknetBodyV1
+from ..classification.darknet import default_cfgs as dark_cfgs
 from ..utils import conv_sequence, load_pretrained_params
 
 __all__ = ['YOLOv1', 'yolov1']
