+++ holocron/models/detection/yolov2.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022, François-Guillaume Fernandez.
+# Copyright (C) 2020-2022, François-Guillaume Fernandez.
 
 # This program is licensed under the Apache License version 2.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0.txt> for full license details.
@@ -14,8 +14,8 @@ from torchvision.ops.misc import FrozenBatchNorm2d
 from holocron.nn import ConcatDownsample2d
 from holocron.nn.init import init_module
 
-from ..darknetv2 import DarknetBodyV2
-from ..darknetv2 import default_cfgs as dark_cfgs
+from ..classification.darknetv2 import DarknetBodyV2
+from ..classification.darknetv2 import default_cfgs as dark_cfgs
 from ..utils import conv_sequence, load_pretrained_params
 from .yolo import _YOLO
